api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "mrkhan-20",email = "taha65641@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.123.8,<0.124.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "psycopg2 (>=2.9.11,<3.0.0)"
]

[tool.poetry]
packages = [{include = "api", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
